Remove support for fuse
This feature adds significant complexity to the library (and testing),
while not seeming terribly useful. End-users can implement it
themselves, and get the timestamp semantics that they need. Fuse also
imposes restrictions on the implementation of the library itself, by
requiring that we wait for *all* senders for a given timestamp before we
send anything on. There's also the issue of what we do when different
updates come into a fuse for the same timestamp, something the current
implementation just ignores.