Enforce sequential timestamps in forwarding mode
This changes the semantics of clocked_dispatch somewhat, but greatly
simplifies implementation. It also arguably makes for a nicer end-user
interface. The dispatcher now assumes that all senders collectively
agree on a single, sequential source of time. Thus, we no longer need to
wait on all senders for any given timestamp. Instead, we can keep track
of the highest sequential timestamp seen thus far, and buffer any
messages that do not have that timestamp + 1. When a message with the
next sequential timestamp is received, all consecutive messages are also
released.